cleanup and modernize django project structure
