+++ .github/workflows/juntagrico-ci.yml
@@ -14,7 +14,7 @@ jobs:
     name: Tests
     runs-on: ubuntu-latest
     env:
-      DJANGO_SETTINGS_MODULE: testsettings
+      DJANGO_SETTINGS_MODULE: settings.dev
       PYTHONPATH: .
       EXCLUDE_TEST: ${{ matrix.exclude-test }}
       COVERAGE_FILE: coverage${{ matrix.exclude-test }}
@@ -77,7 +77,7 @@ jobs:
       run: |
         python -m django makemigrations --noinput
         coverage run -m manage test --exclude-tag=$EXCLUDE_TEST
-        python -m manage test juntagrico.tests.test_deploy --settings="deploytestsettings"
+        python -m manage test juntagrico.tests.test_deploy --settings="settings.deployment"
     - name: format coverage
       if: ${{ matrix.python-version  == '3.11' }}
       run: |
