+++ manage.py
@@ -5,5 +5,5 @@ import sys
 from django.core.management import execute_from_command_line
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "testsettings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings.dev")
     execute_from_command_line(sys.argv)
