+++ urls/dev.py
@@ -17,8 +17,6 @@ from django.conf.urls import include
 from django.contrib import admin
 from django.urls import path
 
-from juntagrico import views
-
 urlpatterns = [
     path('admin/', admin.site.urls),
 
@@ -26,5 +24,4 @@ urlpatterns = [
     path('djrichtextfield/', include('djrichtextfield.urls')),
 
     path('', include('juntagrico.urls')),
-    path('', views.home),
 ]
